Added birthday to baby creation
Dates are weird in firebase.
Needs to be stored as a long.

Program will take user data, put it in a date, convert the date to a long and then push to firebase.

When we access birthDate, it will be as a long so need to convert it back into a date when pulling from firebase.